Refactor: create different header for the project

+++ tests/allignement.c
@@ -14,10 +14,10 @@ UTEST_F(resetMalloc, allignment_is_fine) {
 		addr[a] = mymalloc(3 * sizeof(void *));
 		ASSERT_EQ(0, ((uint64_t)addr[a] % 16));
 		addr[a][0] = mymalloc(TINY_TEST_SIZE);
-		ASSERT_EQ(0, (int)addr[a][0] % 16);
+		ASSERT_EQ(0, (uint64_t)addr[a][0] % 16);
 		addr[a][1] = mymalloc(MEDIUM_TEST_SIZE); 
-		ASSERT_EQ(0, (int)addr[a][1] % 16);
+		ASSERT_EQ(0, (uint64_t)addr[a][1] % 16);
 		addr[a][2] = mymalloc(LARGE_TEST_SIZE);
-		ASSERT_EQ(0, (int)addr[a][2] % 16);
+		ASSERT_EQ(0, (uint64_t)addr[a][2] % 16);
 	}
 }
